# docs/swagger.yaml
basePath: /
definitions:
  entity.Partnership:
    properties:
      created_at:
        type: string
      ended_at:
        type: string
      id:
        type: integer
      status:
        $ref: '#/definitions/entity.Status'
      trainer_id:
        type: integer
      user_id:
        type: integer
    type: object
...
  entity.Role:
    enum:
    - user
    - trainer
    type: string
    x-enum-varnames:
    - UserRole
    - TrainerRole
  entity.Status:
    enum:
    - approved
    - request
    - ended by user
    - ended by trainer
    type: string
    x-enum-varnames:
    - StatusApproved
    - StatusRequest
    - StatusEndedByUser
    - StatusEndedByTrainer
  entity.UpdateWorkout:
    properties:
      date:
        type: string
      description:
        type: string
      title:
        type: string
    type: object
  entity.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password_hash:
        type: string
      role:
        $ref: '#/definitions/entity.Role'
      surname:
        type: string
    required:
    - email
    - name
    - password_hash
    - surname
    type: object
...
  entity.Workout:
    properties:
      date:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      trainer_id:
        type: integer
      user_id:
        type: integer
    required:
    - title
    type: object
...
  handler.errorResponse:
    properties:
      error:
        type: string
    type: object
...
  handler.partnershipIdResponse:
    properties:
      partnership_id:
        type: integer
    type: object
  handler.partnershipsResponse:
    properties:
      partnerships:
        items:
          $ref: '#/definitions/entity.Partnership'
        type: array
    type: object
  handler.requestIdResponse:
    properties:
      request_id:
        type: integer
    type: object
...
  handler.usersResponse:
    properties:
      users:
        items:
          $ref: '#/definitions/entity.User'
        type: array
    type: object
  handler.workoutIdResponse:
    properties:
      workout_id:
        type: integer
    type: object
  handler.workoutsResponse:
    properties:
      workouts:
        items:
          $ref: '#/definitions/entity.Workout'
        type: array
    type: object
host: 143.198.157.158:8001
info:
  contact: {}
  description: API Server for Fitness application
  title: Fitness REST API
  version: "1.0"
paths:
...
  /user:
    get:
      description: get information about yourself
      operationId: get-user-info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user info
      tags:
      - user
  /user/partnership:
    get:
      description: get information about your partnerships
      operationId: get-partnership
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.partnershipsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get partnerships
      tags:
      - user
  /user/partnership/trainer/:id:
    post:
      description: sends request to trainer to become his client
      operationId: send-request
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.requestIdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Send request
      tags:
      - user
    put:
      description: ends partnership with trainer
      operationId: end-partnership-as-user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.partnershipIdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: End partnership
      tags:
      - user
  /user/trainer:
    get:
      description: get information about all trainers
      operationId: get-trainers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.usersResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all trainers
      tags:
      - user
  /user/trainer/:id:
    get:
      description: get information about trainer using id
      operationId: get-trainer-by-id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get trainer
      tags:
      - user
  /user/workout:
    get:
      description: get information about your workouts
      operationId: get-user-workouts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.workoutsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all workouts
      tags:
      - user
    post:
      consumes:
      - application/json
      description: creates workout
      operationId: create-workout-as-user
      parameters:
      - description: workout info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/entity.Workout'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.workoutIdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create workout
      tags:
      - user
  /user/workout/:id:
    delete:
      description: deletes workout
      operationId: delete-workout-user
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete workout
      tags:
      - user
    get:
      description: get information about workout using workout id
      operationId: get-workout-user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Workout'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get workout by id
      tags:
      - user
    put:
      consumes:
      - application/json
      description: updates workout
      operationId: update-workout-user
      parameters:
      - description: update workout info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateWorkout'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.workoutIdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update workout
      tags:
      - user
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"